上划进入多任务，动效降档（分辨率） Co-Authored-By: Agent Signed-off-by: anderxiaml <xiapengcheng4@h-partners.com>

+++ rosen/modules/render_service/core/feature/uifirst/rs_sub_thread_cache.cpp
@@ -345,8 +345,8 @@ void RsSubThreadCache::InitCacheSurface(Drawing::GPUContext* gpuContext,
     if (LIKELY(uniParam) && uniParam->IsUifirstScale()) {
         float uifirstScale = uniParam->GetUiFirstScale();
         RS_TRACE_NAME_FMT("%s uifirstScale called scaleRatio is %f", __func__, uifirstScale);
-        width = width * uifirstScale;
-        height = height * uifirstScale;
+        width = static_cast<int>(width * uifirstScale);
+        height = static_cast<int>(height * uifirstScale);
     }
 
     auto params = static_cast<RSSurfaceRenderParams*>(nodeDrawable->GetUifirstRenderParams().get());
